删减重复代码
